+++ packages/network-controller/jest.config.js
@@ -17,7 +17,7 @@ module.exports = merge(baseConfig, {
   // An object that configures minimum threshold enforcement for coverage results
   coverageThreshold: {
     global: {
-      branches: 88.47,
+      branches: 88.02,
       functions: 97.43,
       lines: 94.47,
       statements: 94.29,